Event interval and placeholder
Adding event-interval between start of OW and first event within. Still need to make sure its only happening with appropriate CMA and fix the javascript

Some changes to placeholder code to allow output on html-plot-number > 1 still not able to embed for > 1 probably wont change unless there is a demand for it.